% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualis_sucupira.R
\name{get_all_tables}
\alias{get_all_tables}
\title{get_all_tables}
\usage{
get_all_tables(sucupira_get, num_tables = 5, event = "2013-2016")
}
\arguments{
\item{sucupira_get}{refere-se o objeto gerado pela função \code{get_sucupira_page}}

\item{num_tables}{refere-se a quantidade das n primeiras tabelas}

\item{event}{refere-se ao evento de classificação dos periódiocos se é do triênio
ou quadriênio}
}
\value{
Retorna uma lista atualizada de N tabelas do site do Sucupira
}
\description{
Obtém as tabelas de qualis dos cursos
}
\examples{
\dontrun{
sucupira_pg <- get_sucupira_page()
qualis.tab <- get_all_tables(sucupira_pg)
print(head(qualis.tab$ARTES))
}
}
