% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualis_functions.R
\name{get_options}
\alias{get_options}
\title{get_options}
\usage{
get_options(request, form = "evento")
}
\arguments{
\item{request}{refere-se o objeto gerado pela função \code{get_sucupira_page}}

\item{form}{refere-se ao nome do formulário do menu presente no HTML do sucupira,
algumas opções são: evento, area e estrato}
}
\value{
Retorna um dataframe contendo o valor dos campos de opções do menu informado.
}
\description{
Obtém a lista de opções de cada menu do formulário do site
do Sucupira
}
\examples{
\dontrun{
sucupira_pg <- get_sucupira_page()
evento_op <- get_options(request = sucupira_pg, form = 'evento')
print(evento_op)
area_op <- get_options(request = sucupira_pg, form = 'area')
print(head(area_op))
}
}
