% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualis_sucupira.R
\name{get_qualis_table}
\alias{get_qualis_table}
\title{get_qualis_table}
\usage{
get_qualis_table(sucupira_get, area = "computa", event = "2013-2016")
}
\arguments{
\item{sucupira_get}{refere-se o objeto gerado pela função \code{get_sucupira_page}}

\item{area}{refere-se a área que se deseja obter a tabela dos periódicos}

\item{event}{refere-se ao evento de classificação dos periódiocos se é do triênio
ou quadriênio}
}
\value{
Retorna um dataframe contendo o valor dos campos de opções do menu informado.
}
\description{
Obtém a tabela com o nome dos periódicos de acordo com a classificação do
evento e a área desejada
}
\examples{
sucupira_pg <- get_sucupira_page()
computacao_tb <- get_qualis_table(sucupira_pg)
print(head(computacao_tb))
estatistica_tb <- get_qualis_table(sucupira_get = sucupira_pg, area = 'estatística')
print(head(estatistica_tb))
}
