% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualis.R
\name{GetColNames}
\alias{GetColNames}
\title{GetColNames}
\usage{
GetColNames(url = "http://qualis.ic.ufmt.br/",
  xpath = "//*[@id=\\"tb-conferencias\\"]")
}
\arguments{
\item{url}{refere-se a url da página do qualis http://qualis.ic.ufmt.br/}

\item{xpath}{se refere a xpath da tabela desejada}
}
\value{
Retorna um vetor contendo os nomes dos campos da tabela
}
\description{
Obtém o nome das colunas presentes na tabela
}
\examples{
conferencia.cols <- QualisR:::GetColNames(xpath = '//*[@id="tb-conferencias"]')
periodicos.cols <- QualisR:::GetColNames(xpath = '//*[@id="tb-periodicos"]')
print(conferencia.cols)
print(periodicos.cols)
}
\keyword{internal}
