% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualis.R
\name{GetJsonPath}
\alias{GetJsonPath}
\title{GetJsonPath}
\usage{
GetJsonPath(url = "http://qualis.ic.ufmt.br/", tab = "#tb-conferencias")
}
\arguments{
\item{url}{refere-se a url da página do qualis http://qualis.ic.ufmt.br/}

\item{tab}{refere-se ao nome da tabela no "View Source" da página que se deseja recuperar}
}
\value{
Retorna o caminho do arquivo json que contém as informações da tabela
}
\description{
Obtém o caminho do json utilizado para preencher a tabela dos qualis
}
\examples{
tab.conf <- "#tb-conferencias"
tab.per <- "#tb-periodicos"
print(QualisR:::GetJsonPath(tab = tab.conf))
print(QualisR:::GetJsonPath(tab = tab.per))
}
\keyword{internal}
