% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualis.R
\name{GetQualisList}
\alias{GetQualisList}
\title{GetQualisList}
\usage{
GetQualisList()
}
\value{
Retorna uma lista atualizada de data frames obtidos a partir do site dos
qualis da UFMT
}
\description{
Atualiza os dados com as informações dos qualis na área de computação de acordo com
as informações contidas em http://qualis.ic.ufmt.br/
}
\examples{
qualis.tabs <- GetQualisList()
print(head(qualis.tabs$conferencias))
}
