% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualis.R
\name{GetTable}
\alias{GetTable}
\title{GetTable}
\usage{
GetTable(url = "http://qualis.ic.ufmt.br/", id_table = "tb-conferencias")
}
\arguments{
\item{url}{refere-se a url da página do qualis http://qualis.ic.ufmt.br/}

\item{id_table}{refere-se ao id tabela no "View Source" da página que se deseja recuperar}
}
\value{
Retorna um data.frame contendo as informações referente a tabela buscada
}
\description{
Obtém a tabela do qualis em forma de data.frame
}
\examples{
tab.conf <- GetTable(id_table = "tb-conferencias")
tab.per <-  GetTable(id_table = "tb-periodicos")
print(head(tab.conf))
print(head(tab.per))
}
