% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualis.R
\docType{data}
\name{qualis}
\alias{qualis}
\title{Qualis das conferências e revistas.}
\format{Uma lista contendo três data frames
\describe{
  \item{conferencias}{Qualis das conferências}
  \item{periodicos}{Qualis dos periódicos}
  \item{outros}{Correlação com outras áreas}
}}
\source{
\url{ http://qualis.ic.ufmt.br/}
}
\usage{
qualis
}
\description{
Contém o conjunto de informações sobre qualis de períodicos e conferências da área da computação
}
\keyword{datasets}
